pyproject.toml
[project]
name = "ct_agent_tg_bot_module"
version = "0.1.0"
description = "Telegram bot service for trade signal intake"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"python-telegram-bot>=20.0",
	"pydantic>=2.0",
	"shared",
	"ct_agent_nlp_parser_module",
	"account_manager",
]

[tool.uv.sources]
shared = { workspace = true }
ct_agent_nlp_parser_module = { workspace = true }
account_manager = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["telegram_bot"]